+++ lib/common/utils.c
@@ -28,13 +28,12 @@
 /*
  *  a queue of nodes
  */
-nodequeue *new_queue(int sz)
-{
+nodequeue *new_queue(size_t sz) {
     nodequeue *q = gv_alloc(sizeof(nodequeue));
 
     if (sz <= 1)
 	sz = 2;
-    q->head = q->tail = q->store = gv_calloc((size_t)sz, sizeof(node_t*));
+    q->head = q->tail = q->store = gv_calloc(sz, sizeof(node_t*));
     q->limit = q->store + sz;
     return q;
 }