+++ CHANGELOG.md
@@ -20,6 +20,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The limit of 5 unique `samehead` or `sametail` values per node has been
   removed. The maximum number of unique `samehead` or `sametail` values is now
   limited only by available memory. #452
+- **Breaking**: The `size` field of the `elist` struct is now a `size_t`.
+- **Breaking**: The `size` field of the `nlist` struct is now a `size_t`.
+- **Breaking**: The `n_nodes` field of the `Agraphinfo_t` struct is now a
+  `size_t`.
 
 ### Fixed